+++ src/components/Stat/index.ts
@@ -1,0 +1,3 @@
+export { Stat } from "./Stat";
+export { StatBand } from "./StatBand";
+export type { StatProps, StatBandProps } from "./Stat.types";